Bug fixes to sorting structure, adding lexicographic sort, adding code to compare sam files

+++ bwa.h
@@ -37,9 +37,9 @@ typedef struct {
     uint16_t sam_size;
     int chr_num;
     int64_t abs_pos;            // Actual position with correction from cigar
-    uint8_t correction;         // Correction applied on cigar. To get sam mapping position from this, add correction to abs_pos
+    int8_t correction;         // Correction applied on cigar. To get sam mapping position from this, add correction to abs_pos
     uint8_t is_rev;
-    uint8_t avg_qual;
+    uint16_t avg_qual;
     uint16_t mate_diff;
     uint16_t flags;
     uint8_t last_entry;
